+++ include/sound/pcm.h
@@ -1500,7 +1500,7 @@ extern const struct snd_pcm_chmap_elem snd_pcm_alt_chmaps[];
 int snd_pcm_add_chmap_ctls(struct snd_pcm *pcm, int stream,
 			   const struct snd_pcm_chmap_elem *chmap,
 			   int max_channels,
-			   unsigned long private_value,
+			   uintptr_t private_value,
 			   struct snd_pcm_chmap **info_ret);
 
 /**
