+++ sound/hda/codecs/realtek/alc268.c
@@ -9,7 +9,7 @@ static int alc268_beep_switch_put(struct snd_kcontrol *kcontrol,
 				  struct snd_ctl_elem_value *ucontrol)
 {
 	struct hda_codec *codec = snd_kcontrol_chip(kcontrol);
-	unsigned long pval;
+	uintptr_t pval;
 	int err;
 
 	guard(mutex)(&codec->control_mutex);
